+++ CHANGELOG.md
@@ -83,6 +83,11 @@ and this project adheres to
   The main effect of this is, when post-styling a table with CSS, cell content
   remains visible. I.e. Z-ordering is more intuitive.
 
+### Fixed
+
+- Nodes with multiple peripheries no longer grow on repeated DOT round trips.
+  #813
+
 ## [15.1.0] – 2026-06-17
 
 ### Added